Move serializer list from utils to core
