+++ packages/core/cache/package.json
@@ -17,5 +17,8 @@
   "dependencies": {
     "@parcel/logger": "^2.0.0-alpha.3.1",
     "@parcel/utils": "^2.0.0-alpha.3.1"
+  },
+  "peerDependencies": {
+    "@parcel/core": "^2.0.0-alpha.3.1"
   }
 }
