+++ packages/core/core/src/index.js
@@ -1,5 +1,15 @@
 // @flow
 
+// Needs to be exported first because of circular imports
+export {
+  registerSerializableClass,
+  unregisterSerializableClass,
+  prepareForSerialization,
+  restoreDeserializedObject,
+  serialize,
+  deserialize,
+} from './serializer';
+
 export {
   default,
   BuildError,
