+++ packages/core/core/src/serializer.js
