+++ packages/core/core/src/utils.js
@@ -3,6 +3,16 @@
 import type {AbortSignal} from 'abortcontroller-polyfill/dist/cjs-ponyfill';
 import type {BundleGroup} from '@parcel/types';
 
+import {registerSerializableClass} from './serializer';
+import AssetGraph from './AssetGraph';
+import BundleGraph from './BundleGraph';
+import Graph from './Graph';
+import ParcelConfig from './ParcelConfig';
+import {RequestGraph} from './RequestTracker';
+import Config from './public/Config';
+// $FlowFixMe this is untyped
+import packageJson from '../package.json';
+
 export function getBundleGroupId(bundleGroup: BundleGroup): string {
   return 'bundle_group:' + bundleGroup.entryAssetId;
 }
@@ -15,4 +25,29 @@ export function assertSignalNotAborted(signal: ?AbortSignal): void {
 
 export class BuildAbortError extends Error {
   name = 'BuildAbortError';
+}
+
+let coreRegistered;
+export function registerCoreWithSerializer() {
+  if (coreRegistered) {
+    return;
+  }
+
+  const packageVersion: mixed = packageJson.version;
+  if (typeof packageVersion !== 'string') {
+    throw new Error('Expected package version to be a string');
+  }
+
+  for (let ctor of [
+    AssetGraph,
+    Config,
+    BundleGraph,
+    Graph,
+    ParcelConfig,
+    RequestGraph,
+  ]) {
+    registerSerializableClass(packageVersion + ':' + ctor.name, ctor);
+  }
+
+  coreRegistered = true;
 }
