+++ packages/core/core/test/serializer.test.js
