+++ packages/core/fs/package.json
@@ -23,5 +23,8 @@
     "ncp": "^2.0.0",
     "nullthrows": "^1.1.1",
     "rimraf": "^2.6.2"
+  },
+  "peerDependencies": {
+    "@parcel/core": "^2.0.0-alpha.3.1"
   }
 }
