+++ packages/core/package-manager/package.json
@@ -25,5 +25,8 @@
     "nullthrows": "^1.1.1",
     "resolve": "^1.12.0",
     "split2": "^3.1.1"
+  },
+  "peerDependencies": {
+    "@parcel/core": "^2.0.0-alpha.3.1"
   }
 }
