+++ packages/core/package-manager/src/Yarn.js
@@ -8,7 +8,7 @@ import logger from '@parcel/logger';
 import split from 'split2';
 import JSONParseStream from './JSONParseStream';
 import promiseFromProcess from './promiseFromProcess';
-import {registerSerializableClass} from '@parcel/utils';
+import {registerSerializableClass} from '@parcel/core';
 // $FlowFixMe
 import pkg from '../package.json';
