+++ packages/core/source-map/package.json
@@ -22,5 +22,8 @@
   },
   "devDependencies": {
     "clone": "^2.1.1"
+  },
+  "peerDependencies": {
+    "@parcel/core": "^2.0.0-alpha.3.1"
   }
 }
